Add files via upload
There are many different types of visualizations that you can use, and the best choice for a particular dataset will depend on the nature of the data and the insights you are trying to convey. Some common types of visualizations for exploring a heart disease dataset include scatter plots, line plots, bar plots, and histograms.